Merged original

+++ src/dos/cdrom.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2002-2020  The DOSBox Team
+ *  Copyright (C) 2002-2021  The DOSBox Team
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -223,13 +223,13 @@ private:
 		TrackFile(uint16_t _chunkSize) : chunkSize(_chunkSize) {}
 	public:
 		virtual          ~TrackFile() = default;
-		virtual bool     read(uint8_t *buffer, int seek, int count) = 0;
-		virtual bool     seek(uint32_t offset) = 0;
+		virtual bool     read(uint8_t *buffer,int64_t seek, int count) = 0;
+		virtual bool     seek(int64_t offset) = 0;
 		virtual uint16_t   decode(uint8_t *buffer) = 0;
 		virtual uint16_t   getEndian() = 0;
 		virtual uint32_t   getRate() = 0;
 		virtual uint8_t    getChannels() = 0;
-		virtual int      getLength() = 0;
+		virtual int64_t    getLength() = 0;
 		virtual void setAudioPosition(uint32_t pos) = 0;
 		const uint16_t chunkSize = 0;
 		uint32_t audio_pos = UINT32_MAX; // last position when playing audio
@@ -245,13 +245,13 @@ private:
 		BinaryFile      (const BinaryFile&) = delete; // prevent copying
 		BinaryFile&     operator= (const BinaryFile&) = delete; // prevent assignment
 
-		bool            read(uint8_t *buffer, int seek, int count);
-		bool            seek(uint32_t offset);
+		bool            read(uint8_t *buffer,int64_t seek, int count);
+		bool            seek(int64_t offset);
 		uint16_t          decode(uint8_t *buffer);
 		uint16_t          getEndian();
 		uint32_t          getRate() { return 44100; }
 		uint8_t           getChannels() { return 2; }
-		int             getLength();
+		int64_t           getLength();
 		void setAudioPosition(uint32_t pos) { audio_pos = pos; }
 	private:
 		std::ifstream   *file;
@@ -266,13 +266,13 @@ private:
 		AudioFile       (const AudioFile&) = delete; // prevent copying
 		AudioFile&      operator= (const AudioFile&) = delete; // prevent assignment
 
-		bool            read(uint8_t *buffer, int seek, int count) { (void)buffer; (void)seek; (void)count; return false; }
-		bool            seek(uint32_t offset);
+		bool            read(uint8_t *buffer,int64_t seek, int count) { (void)buffer; (void)seek; (void)count; return false; }
+		bool            seek(int64_t offset);
 		uint16_t          decode(uint8_t *buffer);
 		uint16_t          getEndian();
 		uint32_t          getRate();
 		uint8_t           getChannels();
-		int             getLength();
+		int64_t           getLength();
         void setAudioPosition(uint32_t pos) { (void)pos;/*unused*/ }
 	private:
 		Sound_Sample    *sample = nullptr;
@@ -287,13 +287,13 @@ private:
         CHDFile(const CHDFile&) = delete;
         CHDFile& operator= (const CHDFile&) = delete;
 
-        bool            read(uint8_t* buffer, int seek, int count);
-        bool            seek(uint32_t offset);
+        bool            read(uint8_t* buffer, int64_t seek, int count);
+        bool            seek(int64_t offset);
         uint16_t        decode(uint8_t* buffer);
         uint16_t        getEndian();
         uint32_t        getRate() { return 44100; }
         uint8_t         getChannels() { return 2; }
-        int             getLength();
+        int64_t         getLength();
         void setAudioPosition(uint32_t pos) { audio_pos = pos; }
         chd_file*       getChd() { return this->chd; }
     private:
